[docs] Move link to text component to encourage usage of A
